+++ src/themes/tailwind.js
@@ -207,7 +207,7 @@ export class tailwindTheme extends AbstractTheme {
     return el
   }
 
-  getFormControl (label, input, description, infoText) {
+  getFormControl (label, input, description, infoText, formName) {
     const group = document.createElement('div')
     group.classList.add('form-group', 'mb-1', 'w-full')
     if (label) {
@@ -236,6 +236,16 @@ export class tailwindTheme extends AbstractTheme {
     }
 
     if (description) group.appendChild(description)
+
+    if (input.tagName.toLowerCase() !== 'div' && input && label && formName) {
+      label.setAttribute('for', formName)
+      input.setAttribute('id', formName)
+    }
+
+    if (input.tagName.toLowerCase() !== 'div' && input && description) {
+      description.setAttribute('id', formName + '-description')
+      input.setAttribute('aria-describedby', formName + '-description')
+    }
 
     return group
   }